requirements.txt
feedparser==6.0.11
httpx==0.27.2
jinja2==3.1.4
python-dotenv==1.0.1
google-generativeai==0.8.3
